Add security to product & profile views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -17,7 +18,7 @@ def all_products(request):
     direction= None
     current_gender = None
 
-# Product filtering
+    # Product filtering
     if request.GET:
 
         # Sorting
@@ -104,9 +105,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """ A view to add a new product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to add products.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -126,8 +131,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_product(request, product_id):
     """ A view to edit an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to add products.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -150,8 +160,13 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_product(request, product_id):
     """ A view to delete an existing product """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission to add products.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product "{product.name}" deleted!')
